+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ASKit Changelog
 ===============
+##0.3
+- Added wrapper around NSUserNotification
+    - Shortcuts for setting params and sending notifications
+    - Only available on OS X 10.8 and above
+
 ##0.2
 - Added ping host checker (ASPing.h)
 - Updated all headers and adjusted license